Refactor font approach
