+++ vite.config.js
@@ -1,12 +1,29 @@
 import { sveltekit } from '@sveltejs/kit/vite';
-import wasm from 'vite-plugin-wasm';
+import fs from 'fs';
 
 /** @type {import('vite').UserConfig} */
 const config = {
-	plugins: [wasm(), sveltekit()],
+	plugins: [sveltekit(), rawFonts(['.ttf'])],
 	optimizeDeps: {
 		exclude: ['@vercel/og']
 	}
 };
+
+// adapted from mattjennings
+// https://github.com/mattjennings/mattjennings.io/blob/master/vite.config.js
+function rawFonts(ext) {
+	return {
+		name: 'vite-plugin-raw-fonts',
+		resolveId(id) {
+			return ext.some((e) => id.endsWith(e)) ? id : null;
+		},
+		transform(code, id) {
+			if (ext.some((e) => id.endsWith(e))) {
+				const buffer = fs.readFileSync(id);
+				return { code: `export default ${JSON.stringify(buffer)}`, map: null };
+			}
+		}
+	};
+}
 
 export default config;
